Refactor all existing CamelBlueprintTestSupport tests to use a common CamelBlueprintTest class.
