+++ core/test-api/pom.xml
@@ -11,6 +11,7 @@
   <name>OpenNMS Core Test APIs</name>
   <packaging>pom</packaging>
   <modules>
+    <module>camel</module>
     <module>db</module>
     <module>dns</module>
     <module>http</module>
