+++ features/amqp/event-receiver/src/test/java/org/opennms/features/amqp/eventreceiver/AMQPEventReceiverBlueprintTest.java
@@ -28,16 +28,15 @@
 
 package org.opennms.features.amqp.eventreceiver;
 
-import org.apache.camel.util.KeyValueHolder;
-
 import java.util.Dictionary;
 import java.util.List;
 import java.util.Map;
 
 import org.apache.camel.BeanInject;
-import org.apache.camel.test.blueprint.CamelBlueprintTestSupport;
+import org.apache.camel.util.KeyValueHolder;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.opennms.core.test.camel.CamelBlueprintTest;
 import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.dao.mock.MockEventIpcManager;
 import org.opennms.netmgt.dao.mock.MockNodeDao;
@@ -56,7 +55,7 @@ import com.google.common.collect.Lists;
  * @author jwhite
  */
 @Ignore
-public class AMQPEventReceiverBlueprintTest extends CamelBlueprintTestSupport {
+public class AMQPEventReceiverBlueprintTest extends CamelBlueprintTest {
 
     @BeanInject
     protected EventIpcManager eventIpcManager;
