+++ features/discovery/src/test/java/org/opennms/netmgt/discovery/DiscovererBlueprintIT.java
@@ -34,11 +34,11 @@ import java.util.Properties;
 
 import org.apache.camel.Component;
 import org.apache.camel.component.seda.SedaComponent;
-import org.apache.camel.test.blueprint.CamelBlueprintTestSupport;
 import org.apache.camel.util.KeyValueHolder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opennms.core.test.OpenNMSJUnit4ClassRunner;
+import org.opennms.core.test.camel.CamelBlueprintTest;
 import org.opennms.netmgt.config.DiscoveryConfigFactory;
 import org.opennms.netmgt.config.api.DiscoveryConfigurationFactory;
 import org.opennms.netmgt.config.discovery.DiscoveryConfiguration;
@@ -49,50 +49,15 @@ import org.opennms.netmgt.events.api.EventForwarder;
 import org.opennms.netmgt.events.api.EventIpcManager;
 import org.opennms.netmgt.icmp.Pinger;
 import org.opennms.netmgt.model.OnmsDistPoller;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.test.context.ContextConfiguration;
 
 @RunWith( OpenNMSJUnit4ClassRunner.class )
 @ContextConfiguration( locations = { "classpath:/META-INF/opennms/emptyContext.xml" } )
-public class DiscovererBlueprintIT extends CamelBlueprintTestSupport {
+public class DiscovererBlueprintIT extends CamelBlueprintTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger(DiscoveryBlueprintIT.class );
     private static final MockEventIpcManager IPC_MANAGER_INSTANCE = new MockEventIpcManager();
 
     private static final String LOCATION = "RDU";
-
-    /**
-     * Use Aries Blueprint synchronous mode to avoid a blueprint deadlock bug.
-     * 
-     * @see https://issues.apache.org/jira/browse/ARIES-1051
-     * @see https://access.redhat.com/site/solutions/640943
-     */
-    @Override
-    public void doPreSetup() throws Exception
-    {
-        System.setProperty( "org.apache.aries.blueprint.synchronous", Boolean.TRUE.toString() );
-        System.setProperty( "de.kalpatec.pojosr.framework.events.sync", Boolean.TRUE.toString() );
-    }
-
-    @Override
-    public boolean isUseAdviceWith()
-    {
-        return true;
-    }
-
-    @Override
-    public boolean isUseDebugger()
-    {
-        // must enable debugger
-        return true;
-    }
-
-    @Override
-    public String isMockEndpoints()
-    {
-        return "*";
-    }
 
     /**
      * Register a mock OSGi {@link SchedulerService} so that we can make sure that the scheduler
@@ -100,8 +65,7 @@ public class DiscovererBlueprintIT extends CamelBlueprintTestSupport {
      */
     @SuppressWarnings( "rawtypes" )
     @Override
-    protected void addServicesOnStartup( Map<String, KeyValueHolder<Object, Dictionary>> services )
-    {
+    protected void addServicesOnStartup( Map<String, KeyValueHolder<Object, Dictionary>> services ) {
         services.put( Pinger.class.getName(), new KeyValueHolder<Object, Dictionary>(new TestPinger(), new Properties()));
 
         services.put( EventForwarder.class.getName(),
@@ -140,12 +104,12 @@ public class DiscovererBlueprintIT extends CamelBlueprintTestSupport {
 
     // The location of our Blueprint XML file to be used for testing
     @Override
-    protected String getBlueprintDescriptor()
-    {
+    protected String getBlueprintDescriptor() {
         return "file:blueprint-discoverer.xml";
     }
 
     @Test
     public void testDiscoverer() throws Exception {
+        // pass
     }
 }
