+++ features/discovery/src/test/java/org/opennms/netmgt/discovery/DiscoveryBlueprintDistPollerDaoIT.java
@@ -32,54 +32,21 @@ import java.util.Dictionary;
 import java.util.Map;
 import java.util.Properties;
 
-import org.apache.camel.test.blueprint.CamelBlueprintTestSupport;
 import org.apache.camel.util.KeyValueHolder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opennms.core.test.OpenNMSJUnit4ClassRunner;
+import org.opennms.core.test.camel.CamelBlueprintTest;
 import org.opennms.minion.core.api.MinionIdentity;
 import org.opennms.netmgt.dao.api.DistPollerDao;
 import org.opennms.netmgt.model.OnmsDistPoller;
 import org.opennms.netmgt.model.OnmsMonitoringSystem;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.test.context.ContextConfiguration;
 
 @RunWith( OpenNMSJUnit4ClassRunner.class )
 @ContextConfiguration( locations = { "classpath:/META-INF/opennms/emptyContext.xml" } )
-public class DiscoveryBlueprintDistPollerDaoIT extends CamelBlueprintTestSupport
-{
-    private static final Logger LOG = LoggerFactory.getLogger(DiscoveryBlueprintDistPollerDaoIT.class);
-
+public class DiscoveryBlueprintDistPollerDaoIT extends CamelBlueprintTest {
     private static final String LOCATION = "TEST_LOCATION";
-
-    /**
-     * Use Aries Blueprint synchronous mode to avoid a blueprint deadlock bug.
-     * 
-     * @see https://issues.apache.org/jira/browse/ARIES-1051
-     * @see https://access.redhat.com/site/solutions/640943
-     */
-    @Override
-    public void doPreSetup() throws Exception {
-        System.setProperty( "org.apache.aries.blueprint.synchronous", Boolean.TRUE.toString() );
-        System.setProperty( "de.kalpatec.pojosr.framework.events.sync", Boolean.TRUE.toString() );
-    }
-
-    @Override
-    public boolean isUseAdviceWith() {
-        return true;
-    }
-
-    @Override
-    public boolean isUseDebugger() {
-        // must enable debugger
-        return true;
-    }
-
-    @Override
-    public String isMockEndpoints() {
-        return "*";
-    }
 
     @SuppressWarnings( "rawtypes" )
     @Override
